refactoring: saparated methods in different folders
