+++ .flake8
@@ -1,4 +1,4 @@
 [flake8]
 max-line-length = 120
-exclude = .git,__pycache__,venv, migrations
+exclude = .git,__pycache__,venv, migrations, .\db\models\__init__.py,
 ignore = E203,W503, E722
