+++ db/models/__init__.py
@@ -1,3 +1,3 @@
-from .companies import Company
-from .association import company_activity_association
 from .activities import Activity
+from .association import company_activity_association
+from .companies import Company
